+++ src/query/genetic-analysts/genetic-analyst-qualification.ts
